DocuShield-AI (1)/requirements.txt
flask
werkzeug
PyPDF2
python-docx
python-pptx
openpyxl
odfpy
